Move Params to config
